fix: switching to AppDelegate
