+++ SNUTT-2022/SNUTT/Views/SNUTTApp.swift
@@ -1,68 +1,1 @@
-//
-//  SNUTTApp.swift
-//  SNUTT
-//
-//  Created by Jinsup Keum on 2022/03/19.
-//
 
-import SwiftUI
-import UIKit
-
-/// Wrapper struct that supports both iOS 13 and iOS 14 (and above).
-@main
-struct SNUTTAppWrapper {
-    static func main() {
-        if #available(iOS 14.0, *) {
-            SNUTTApp.main()
-        } else {
-            /// In iOS 13, use `SceneDelegate` instead of `@main`.
-            UIApplicationMain(CommandLine.argc, CommandLine.unsafeArgv, nil, NSStringFromClass(SceneDelegate.self))
-        }
-    }
-}
-
-@available(iOS 14.0, *)
-struct SNUTTApp: App {
-    var body: some Scene {
-        WindowGroup {
-            ContentView()
-        }
-    }
-}
-
-// MARK: Polyfills for iOS 13.
-
-class SceneDelegate: UIResponder, UIWindowSceneDelegate {
-    var window: UIWindow?
-
-    func scene(_ scene: UIScene, willConnectTo _: UISceneSession, options _: UIScene.ConnectionOptions) {
-        let contentView = ContentView()
-
-        // Use a UIHostingController as window root view controller.
-        if let windowScene = scene as? UIWindowScene {
-            let window = UIWindow(windowScene: windowScene)
-            window.rootViewController = UIHostingController(rootView: contentView)
-            self.window = window
-            window.makeKeyAndVisible()
-        }
-    }
-}
-
-// class AppDelegate: UIResponder, UIApplicationDelegate {
-//    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
-//        // Override point for customization after application launch.
-//        return true
-//    }
-//
-//    func application(_ application: UIApplication, configurationForConnecting connectingSceneSession: UISceneSession, options: UIScene.ConnectionOptions) -> UISceneConfiguration {
-//        // Called when a new scene session is being created.
-//        // Use this method to select a configuration to create the new scene with.
-//        return UISceneConfiguration(name: "Default Configuration", sessionRole: connectingSceneSession.role)
-//    }
-//
-//    func application(_ application: UIApplication, didDiscardSceneSessions sceneSessions: Set<UISceneSession>) {
-//        // Called when the user discards a scene session.
-//        // If any sessions were discarded while the application was not running, this will be called shortly after application:didFinishLaunchingWithOptions.
-//        // Use this method to release any resources that were specific to the discarded scenes, as they will not return.
-//    }
-// }
